Add files via upload
I added the collisable class, and added the table and the key. I made some changes to the animated sprite examples. so check that too. I might have changed some things on the others and if there are errors I will show you what you need to add on the other docs.